pyproject.toml
[project]
name = "nanobot-feishu-tools"
version = "0.1.0"
description = "Feishu/Lark tools plugin for nanobot (doc, wiki, bitable, task)"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "nanobot-ai>=0.1.4",
    "lark-oapi>=1.5.0,<2.0.0",
]

[project.entry-points."nanobot.tools"]
feishu = "nanobot_feishu_tools:register_tools"

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nanobot_feishu_tools"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
